backend/requirements.txt
asgiref==3.2.10
psycopg2-binary==2.8.6
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
djoser==2.1.0
requests==2.26.0
django-colorfield==0.7.2
gunicorn==20.0.4
django-filter==2.4.0
drf-extra-fields==3.4.0
django-cors-headers==3.11.0